backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles==23.2.1",
    "aiosmtplib==3.0.1",
    "alembic==1.12.1",
    "apscheduler==3.10.4",
    "asyncpg==0.29.0",
    "black==23.11.0",
    "celery==5.3.4",
    "email-validator==2.1.0",
    "fastapi==0.104.1",
    "flake8==6.1.0",
    "google-generativeai==0.3.2",
    "httpx==0.25.2",
    "mypy==1.7.1",
    "openpyxl==3.1.2",
    "pandas==2.1.3",
    "passlib[bcrypt]==1.7.4",
    "phonenumbers==8.13.25",
    "pydantic[email]==2.5.0",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-mock==3.12.0",
    "python-dateutil==2.8.2",
    "python-dotenv==1.0.0",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "python-telegram-bot==20.7",
    "pytz==2023.3",
    "structlog==23.2.0",
    "uvicorn[standard]==0.24.0",
]
